# docs/openapi.yaml
openapi: 3.1.0
info:
  title: GVoice API
  version: 0.1.0
servers:
  - url: http://localhost:4000
paths:
...
  /workspaces:
    get:
      summary: List user workspaces
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: OK }
    post:
      summary: Create workspace
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, slug]
              properties:
                name: { type: string }
                slug: { type: string }
      responses:
        '201': { description: Created }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT